requirements.txt
# Core LangChain
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-text-splitters==0.3.0

# Vector Store
faiss-cpu==1.9.0

# Embeddings
sentence-transformers==3.3.0
langchain-huggingface==0.1.0

# Google Gemini - FIX: Version tương thích
langchain-google-genai==2.0.0
google-generativeai>=0.7.0,<0.8.0

# UI
streamlit==1.40.0

# Utilities
python-dotenv==1.0.0
jsonschema==4.23.0